[DEBT-043-D1] Split cli/market.py → cli/market/ package
- Converted 788-line cli/market.py into cli/market/ package with 8 files
- Split by command: get.py, orderbook.py, liquidity.py, history.py, list.py, search.py
- Extracted shared helpers to _helpers.py (normalize_market_list_status, optional_lower)
- Created __init__.py with app registration and backwards-compatible exports
- All new files ≤400 lines (largest: list.py at 276 lines)
- Quality gates: pre-commit, mypy, pytest (1052 passed)
- Acceptance criteria: Phase D1 complete
