+++ src/kalshi_research/cli/market/orderbook.py
@@ -1,0 +1,52 @@
+"""Market orderbook command - fetch orderbook for a market."""
+
+from __future__ import annotations
+
+from typing import Annotated
+
+import typer
+from rich.table import Table
+
+from kalshi_research.cli.utils import console, exit_kalshi_api_error, run_async
+
+
+def market_orderbook(
+    ticker: Annotated[str, typer.Argument(help="Market ticker to fetch orderbook for.")],
+    depth: Annotated[int, typer.Option("--depth", "-d", help="Orderbook depth.")] = 5,
+) -> None:
+    """Fetch orderbook for a market."""
+    from kalshi_research.cli.client_factory import public_client
+
+    async def _orderbook() -> None:
+        from kalshi_research.api.exceptions import KalshiAPIError
+
+        async with public_client() as client:
+            try:
+                orderbook = await client.get_orderbook(ticker, depth=depth)
+            except KalshiAPIError as e:
+                exit_kalshi_api_error(e)
+            except Exception as e:
+                console.print(f"[red]Error:[/red] {e}")
+                raise typer.Exit(1) from None
+
+        table = Table(title=f"Orderbook: {ticker}")
+        table.add_column("YES Bids", style="green")
+        table.add_column("NO Bids", style="red")
+
+        yes_bids = orderbook.yes_levels
+        no_bids = orderbook.no_levels
+        max_len = max(len(yes_bids), len(no_bids))
+
+        for i in range(max_len):
+            yes_str = f"{yes_bids[i][0]}¢ x {yes_bids[i][1]}" if i < len(yes_bids) else ""
+            no_str = f"{no_bids[i][0]}¢ x {no_bids[i][1]}" if i < len(no_bids) else ""
+            table.add_row(yes_str, no_str)
+
+        console.print(table)
+
+        if orderbook.spread is not None:
+            console.print(f"\nSpread: {orderbook.spread}¢")
+        if orderbook.midpoint is not None:
+            console.print(f"Midpoint: {orderbook.midpoint:.1f}¢")
+
+    run_async(_orderbook())
